Implemented: Map area select, remove and update.
It is now possible to select an area by activating the polygon icon on the map and double clicking on the map to create a vertex

It is possible to remove the vertices from the created polygon by selecting the vertex and using the trash icon.

The vertices can be thought by clicking and dragging.